Extract cap-floor FpML mapping
